Various fixes

+++ src/main/scala/org/edla/tmdb/api/TmdbApi.scala
@@ -3,7 +3,7 @@ package org.edla.tmdb.api
 import scala.concurrent.Future
 
 trait TmdbApi {
-  def getMovie(id: Int): Future[Movie]
+  def getMovie(id: Long): Future[Movie]
   def getConfiguration(): Future[Configuration]
   def getToken(): Future[AuthenticateResult]
   def searchMovie(query: String): Future[Results]
